Updated Variables - Magic Numbers reduced
I've gone through and updated majority of the variables and made #define's. 

The code is now a lot more readable and compact, however the display / read in code needs to be fixed.

I've made a mistake with the commandInsert function, this only is called by the user and not done automatically - this is why I need to do a call for boolean insert. Need to fix this part of the code this week and enable variable checking for the ID, Name and Telephone (Mostly completed already)